+++ packages/terra-tabs/src/_TabMenu.jsx
@@ -4,7 +4,6 @@ import classNames from 'classnames/bind';
 import Menu from 'terra-menu';
 import IconCaretDown from 'terra-icon/lib/icon/IconCaretDown';
 import { KEY_SPACE, KEY_RETURN } from 'keycode-js';
-import { injectIntl, intlShape } from 'react-intl';
 import { handleArrows } from './_TabUtils';
 import styles from './Tabs.module.scss';
 
@@ -15,20 +14,25 @@ const propTypes = {
    * Tabs that should be displayed collapsed as selectable menu items.
    */
   children: PropTypes.node,
-  /**
-   * @private
-   * intl object programmatically imported through injectIntl from react-intl.
-   * */
-  intl: intlShape.isRequired,
+
   /**
    * Ref callback for menu toggle.
    */
   refCallback: PropTypes.func,
 };
 
+const contextTypes = {
+  /* eslint-disable consistent-return */
+  intl: (context) => {
+    if (context.intl === undefined) {
+      return new Error('Component is internationalized, and must be wrapped in terra-base');
+    }
+  },
+};
+
 class TabMenu extends React.Component {
-  constructor(props) {
-    super(props);
+  constructor(props, context) {
+    super(props, context);
     this.handleOnRequestClose = this.handleOnRequestClose.bind(this);
     this.handleOnClick = this.handleOnClick.bind(this);
     this.handleOnKeyDown = this.handleOnKeyDown.bind(this);
@@ -83,8 +87,9 @@ class TabMenu extends React.Component {
   }
 
   render() {
+    const { intl } = this.context;
     const menuItems = [];
-    let menuToggleText = this.props.intl.formatMessage({ id: 'Terra.tabs.more' });
+    let menuToggleText = intl.formatMessage({ id: 'Terra.tabs.more' });
     let menuActive = false;
 
     React.Children.forEach(this.props.children, (child) => {
@@ -134,6 +139,7 @@ class TabMenu extends React.Component {
   }
 }
 
+TabMenu.contextTypes = contextTypes;
 TabMenu.propTypes = propTypes;
 
-export default injectIntl(TabMenu);
+export default TabMenu;